Changed Float16 to Float32 for age for polars stability

+++ bonsai/functional/create_data.py
@@ -61,7 +61,7 @@ def process_split(
         tokenized = tokenized.select(
             pl.col("subject_id").cast(pl.Int64),
             pl.col("code").cast(pl.Int64),
-            pl.col("age").cast(pl.Float16),
+            pl.col("age").cast(pl.Float32),
             pl.col("abspos").cast(pl.Float32),
             pl.col("segment").cast(pl.Int32),
         )
